refactor(charts): extract reference-line and axis-label rendering into ChartReferenceLayerMixin
